ajuste no pacote

+++ src/api-gateways/AcademyIO.Bff/Models/LessonViewModel.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string Subject { get; set; }
-        public double TotalHours { get; set; }
+        public decimal TotalHours { get; set; }
         public Guid CourseId { get; set; }
     }
 }
