+++ src/services/AcademyIO.Courses.API/Application/Commands/AddCourseCommand.cs
@@ -8,9 +8,9 @@ namespace AcademyIO.Courses.API.Application.Commands
         public string Name { get; set; }
         public string Description { get; set; }
         public Guid UserCreationId { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
-        public AddCourseCommand(string name, string description, Guid userCreationId, double price)
+        public AddCourseCommand(string name, string description, Guid userCreationId, decimal price)
         {
             this.Name = name;
             this.Description = description;
