+++ src/services/AcademyIO.Courses.API/Models/Lesson.cs
@@ -2,11 +2,11 @@
 
 namespace AcademyIO.Courses.API.Models
 {
-    public class Lesson(string name, string subject, double totalHours, Guid courseId) : Entity, IAggregateRoot
+    public class Lesson(string name, string subject, decimal totalHours, Guid courseId) : Entity, IAggregateRoot
     {
         public string Name { get; set; } = name;
         public string Subject { get; set; } = subject;
-        public double TotalHours { get; set; } = totalHours;
+        public decimal TotalHours { get; set; } = totalHours;
         public Guid CourseId { get; set; } = courseId;
     }
 }
