+++ src/services/AcademyIO.Payments.API/AntiCorruption/IPayPalGateway.cs
@@ -6,5 +6,5 @@ public interface IPayPalGateway
 {
     string GetPayPalServiceKey(string apiKey, string encriptionKey);
     string GetCardHashKey(string serviceKey, string cartaoCredito);
-    Transaction CommitTransaction(string cardHashKey, string orderId, double amount);
+    Transaction CommitTransaction(string cardHashKey, string orderId, decimal amount);
 }
